refactor(apps/web): 디렉토리 경로를 단수로 변경
